@@ -1,5 +1,7 @@
-import { createMuiTheme } from '@material-ui/core/styles';
-
-const theme = createMuiTheme({});
+import { createMuiTheme } from "@material-ui/core/styles";
+const theme = createMuiTheme({
+  overrides: {
+  }
+});
 
 export default theme;
