@@ -1,34 +1,28 @@
-import { CssBaseline } from '@material-ui/core';
-import { ThemeProvider } from '@material-ui/styles';
-import React from 'react';
-import 'react-perfect-scrollbar/dist/css/styles.css';
-import 'react-toastify/dist/ReactToastify.css';
-import { Provider } from 'react-redux';
-import AppRouter from '../../commons/Router';
-import theme from '../../commons/Theme';
-import configureStore from '../../redux/configureStore';
-import { ToastContainer } from 'react-toastify';
+import React from "react";
+import PropTypes from "prop-types";
+import configureStore from "../../redux/configureStore";
+import { Provider } from "react-redux";
+import { CssBaseline, ThemeProvider } from "@material-ui/core";
+import theme from "../../commons/Theme";
+import HomePage from "../HomePage/index";
+import CustomModal from "../../components/CustomModal";
+import SigninPage from '../SigninPage/index';
 
 const store = configureStore();
 
-class App extends React.Component {
-  state = {
-    authed: true,
-    loading: false,
-  };
-
-  render() {
-    return (
-      <Provider store={store}>
-        <ThemeProvider theme={theme}>
-          <CssBaseline>
-            <AppRouter />
-            <ToastContainer />
-          </CssBaseline>
-        </ThemeProvider>
-      </Provider>
-    );
-  }
+function App(props) {
+  return (
+    <Provider store={store}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline>
+          <SigninPage />
+          <CustomModal />
+        </CssBaseline>
+      </ThemeProvider>
+    </Provider>
+  );
 }
+
+App.propTypes = {};
 
 export default App;
