@@ -1,25 +1,25 @@
-import { applyMiddleware, compose, createStore } from 'redux';
-import createSagaMiddleware from 'redux-saga';
-import rootReducer from '../reducers';
-import saga from '../sagas';
+import { applyMiddleware, compose, createStore } from "redux";
+import createSagaMiddleware from "redux-saga";
+import rootReducer from "../reducers";
+import saga from "../sagas";
 
 const composeEnhancers =
-  process.env.NODE_ENV !== 'production' &&
-  typeof window === 'object' &&
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
-        shouldHotReload: false,
-      })
-    : compose;
+	process.env.NODE_ENV !== "production" &&
+	typeof window === "object" &&
+	window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+		? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
+				shouldHotReload: false,
+		  })
+		: compose;
 
 const sagaMiddleware = createSagaMiddleware();
 
 const configureStore = () => {
-  const middlewares = [sagaMiddleware];
-  const enhancers = [applyMiddleware(...middlewares)];
-  const store = createStore(rootReducer, composeEnhancers(...enhancers));
-  sagaMiddleware.run(saga);
-  return store;
+	const middlewares = [sagaMiddleware];
+	const enhancers = [applyMiddleware(...middlewares)];
+	const store = createStore(rootReducer, composeEnhancers(...enhancers));
+	sagaMiddleware.run(saga);
+	return store;
 };
 
 export default configureStore;
