@@ -1,11 +1,8 @@
 import { all } from 'redux-saga/effects';
 
-import card from './card';
-import auth from './auth';
-import list from './list';
-
 function* rootSaga() {
-  yield all([...list, ...card, ...auth]);
+	yield all([]);
+	console.log('Root saga');
 }
 
 export default rootSaga;
